Drone software/requirements.txt
numpy
opencv-python
torch
torchvision
ultralytics
mavsdk
gstreamer-python
flask
RPi.GPIO
pyserial
dronekit
pymavlink
